feat: migrate sync onboarding companion seed step to new arch
chore: CR updates

chore: sonar fix:
