+++ .changeset/new-shrimps-provide.md
@@ -1,0 +1,5 @@
+---
+"ledger-live-desktop": minor
+---
+
+Migrate sync onboarding companion seed step to new arch
