+++ apps/ledger-live-desktop/src/config/urls.ts
@@ -202,6 +202,7 @@ export const urls = {
     "https://shop.ledger.com?utm_source=live&utm_medium=draw&utm_campaign=ledger_sync_lns_uncompatible&utm_content=to_shop",
   learnMoreLedgerSync:
     "https://www.ledger.com/blog-ledger-sync-synchronize-your-crypto-accounts-effortless-private-and-secure",
+  charonLearnMore: "https://shop.ledger.com/products/ledger-recovery-key",
 
   // Node errors
   txBroadcastErrors: {
