+++ apps/ledger-live-desktop/src/newArch/features/Onboarding/components/StepText/index.tsx
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+import { VerticalTimeline } from "@ledgerhq/react-ui";
+
+const StepText = styled(VerticalTimeline.BodyText)`
+  white-space: pre-wrap;
+`;
+
+export default StepText;
