+++ apps/ledger-live-desktop/src/newArch/features/Onboarding/screens/SyncOnboardingCompanion/__tests__/NewSeedPanel.test.tsx
@@ -1,6 +1,3 @@
-/**
- * @jest-environment jsdom
- */
 import React from "react";
 import { render, screen } from "tests/testSetup";
