+++ apps/ledger-live-desktop/src/renderer/analytics/hooks/useTrackOnboardingFlow.ts
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 import { Device } from "@ledgerhq/types-devices";
-import { SeedPathStatus } from "~/renderer/components/SyncOnboarding/Manual/SyncOnboardingCompanion/SeedStep";
+import { SeedPathStatus } from "LLD/features/Onboarding/screens/SyncOnboardingCompanion/types";
 import { track } from "../segment";
 import { CONNECTION_TYPES, HOOKS_TRACKING_LOCATIONS } from "./variables";
